Few changes to event handling
* Add in event queue types (just a radix heap for now)
* Change how the damage event works

+++ crates/xivc-core/src/util.rs
@@ -8,13 +8,13 @@ use serde::{Deserialize, Serialize};
 /// A utility function that returns the potency of an action
 /// depending on if it was comboed into and if it hit it's positional.
 pub const fn combo_pos_pot(
-    base: u16,
-    if_pos: u16,
-    if_combo: u16,
-    if_both: u16,
+    base: u64,
+    if_pos: u64,
+    if_combo: u64,
+    if_both: u64,
     combo: bool,
     pos: bool,
-) -> u16 {
+) -> u64 {
     match (combo, pos) {
         (false, false) => base,
         (false, true) => if_pos,
@@ -25,7 +25,7 @@ pub const fn combo_pos_pot(
 
 /// A utility function that returns the potency of an action
 /// depending on if it was comboed into.
-pub const fn combo_pot(base: u16, if_combo: u16, combo: bool) -> u16 {
+pub const fn combo_pot(base: u64, if_combo: u64, combo: bool) -> u64 {
     if combo {
         if_combo
     } else {
@@ -35,7 +35,7 @@ pub const fn combo_pot(base: u16, if_combo: u16, combo: bool) -> u16 {
 
 /// A utility function that returns the potency of an action
 /// depending on if it hit it's positional.
-pub const fn pos_pot(base: u16, if_pos: u16, pos: bool) -> u16 {
+pub const fn pos_pot(base: u64, if_pos: u64, pos: bool) -> u64 {
     if pos {
         if_pos
     } else {
@@ -323,29 +323,34 @@ impl<const MAX: u8> ops::Deref for GaugeU8<MAX> {
 /// #     ActionTargetting,
 /// #     Faction,
 /// #     DamageEventExt,
-/// #     DamageEvent,
 /// #     Actor,
 /// # };
 /// # use xivc_core::timing::{EventCascade};
 /// # use xivc_core::need_target;
+/// # use xivc_core::enums::DamageInstance;
 /// # fn example(world: &impl World, event_sink: &mut impl EventProxy) {
 /// # let src = world.actor(ActorId(0)).unwrap();
+/// // Constants like these are recommended to reduce boilerplate.
 /// const TARGET_CIRCLE: ActionTargetting = ActionTargetting::target_circle(5, 25);
 /// const MELEE: ActionTargetting = ActionTargetting::single(3);
-/// // ...
+/// 
+/// // A closure like this is also recommended to reduce boilerplate.
 /// let target_enemy = |t: ActionTargetting| {
 ///     src.actors_for_action(Some(Faction::Enemy), t).map(|a| a.id())
 /// };
-/// // ...
+/// 
+/// // Deal damage to targets in a circle with a radius of 5y and a range of 25y.
+/// // This aoe will have damage falloff.
 /// let (first, other) = need_target!(target_enemy(TARGET_CIRCLE), event_sink, aoe);
 /// let mut cascade = EventCascade::new(600, 1);
-/// event_sink.damage(DamageEvent::new(1000, first).slashing(), cascade.next());
+/// event_sink.damage(src, DamageInstance::new(1000).slashing(), first, cascade.next());
 /// for target in other {
-///     event_sink.damage(DamageEvent::new(500, target).slashing(), cascade.next());
+///     event_sink.damage(src, DamageInstance::new(500).slashing(), target, cascade.next());
 /// }
-/// // ...
+/// 
+/// // Deal damage to a single target within a range of 3y.
 /// let target = need_target!(target_enemy(MELEE).next(), event_sink);
-/// event_sink.damage(DamageEvent::new(350, target).slashing(), 400);
+/// event_sink.damage(src, DamageInstance::new(350).slashing(), target, 400);
 /// # }
 /// ```
 ///
